tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.roundrect.zero.6.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.roundrect.zero.6</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.roundrect.zero.6</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.strokeStyle = '#f00';
  ctx.lineJoin = 'miter';
  ctx.miterLimit = 1.5;
  ctx.lineWidth = 200;
  ctx.beginPath();
  ctx.roundRect(100, 25, 1000, 0, [0]);
  ctx.stroke();
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
